requirements_dev.txt
-e .
paramiko
factory-boy
django-debug-toolbar
django-extensions
Werkzeug
